Refactor import paths for RoleAdmin and restore enum definitions
