+++ src/app/admin/quan-li-nguoi-dung/page.tsx
@@ -10,7 +10,7 @@ import { UserInfor } from "@/dtos/auth/auth.dto";
 import { UserRequestUpdateUsser } from "@/dtos/auth/auth.request.dto";
 import { UserResponseGetItem } from "@/dtos/auth/auth.response.dto";
 import { UserContractItem } from "@/dtos/quan-li-nguoi-dung/contracts/contract.dto";
-import { RoleAdmin } from "@/model/enum";
+import { RoleAdmin } from "@/models/enum";
 import QuanLyHopDongServices from "@/services/admin/quan-li-nguoi-dung/quan-li-hop-dong.service";
 import QlNguoiDungServices from "@/services/admin/quan-li-nguoi-dung/quan-li-nguoi-dung.service";
 import AuthServices from "@/services/auth/api.service";
