+++ src/app/hr/quan-li-nhan-vien/quan-li-nhan-vien.tsx
@@ -1,5 +1,5 @@
 import FormModal from "@/components/basicUI/FormModal";
-import { RoleAdmin } from "@/model/enum";
+import { RoleAdmin } from "@/models/enum";
 import {
   Avatar,
   Col,
